requirements.txt
numpy==1.23
pandas==1.4
matplotlib==3.5
pillow==9.2
# tensorflow==2.9.1 # cannot be installed with pip -r
#scipy==1.9
#scikit-learn==1.1
#seaborn==0.11
#opencv-python==4.6
